backend/requirements.txt
openai>=0.27.0
pandas>=1.5.0
python-multipart>=0.0.5
Flask>=2.3.0
Flask-CORS>=4.0.0
itsdangerous>=2.1.0
scikit-learn>=1.3.0
joblib>=1.3.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
scipy>=1.11.0
python-dateutil>=2.8.0
xgboost>=2.0.0
lightgbm>=4.0.0
shap>=0.42.0
tenacity>=8.2.0
firebase-admin>=7.0.0
bcrypt>=4.0.0
gunicorn>=21.2.0
